# CheckCheck/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CheckCheck REST API",
    "version": "0.1.dev47+ga91abbb.d20260618"
  },
  "paths": {
...
    "/api/checklist/{checklist_id}/item/{checklist_item_id}/state": {
      "get": {
        "tags": [
          "Checklist Items State"
        ],
        "summary": "Get Checklist Item Checked State",
        "description": "Get the checked value for a specific checklist item",
        "operationId": "get_checklist_item_checked_state_api_checklist__checklist_id__item__checklist_item_id__state_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "checklist_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Checklist Item Id"
            }
          },
          {
            "name": "checklist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Checklist Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckListItemStateWithoutChecklistID"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Checklist Items State"
        ],
        "summary": "Set Checklist Item Checked State",
        "description": "Set the chacked value for a specific checklist item",
        "operationId": "set_checklist_item_checked_state_api_checklist__checklist_id__item__checklist_item_id__state_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "checklist_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Checklist Item Id"
            }
          },
          {
            "name": "checklist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Checklist Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckListItemStateUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckListItemStateWithoutChecklistID"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckListItemStateUpdate": {
        "properties": {
          "checked": {
            "type": "boolean",
            "title": "Checked",
            "default": false
          }
        },
        "type": "object",
        "title": "CheckListItemStateUpdate"
      },
      "CheckListItemStateWithoutChecklistID": {
        "properties": {
          "checked": {
            "type": "boolean",
            "title": "Checked"
          }
        },
        "type": "object",
        "required": [
          "checked"
        ],
        "title": "CheckListItemStateWithoutChecklistID"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}